% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addGoodCalib}
\alias{addGoodCalib}
\title{Good Calibration}
\usage{
addGoodCalib(pred, y, plot.values=TRUE, ...)
}
\arguments{
\item{pred}{predicted probabilities}

\item{y}{response}

\item{plot.values}{wether the value should be plotted}

\item{...}{currently not used}
}
\description{
Displays integrals that refer to good calibration
}
\author{
Giuseppe Casalicchio
}

