% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addPEV}
\alias{addPEV}
\title{Highlights Area that is used to compute the PEV}
\usage{
addPEV(obj, plot.values = TRUE, ...)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{plot.values}{[\code{logical(1)}]\cr
Whether values for the corresponding measure should be added to the plot?
Default is \code{FALSE}.}

\item{...}{passed to abline}
}
\description{
Highlights Area that is used to compute the PEV
}

