% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addPrevalence}
\alias{addPrevalence}
\title{Prevalence}
\usage{
addPrevalence(pred, y, col="gray", plot.values=TRUE, ...)
}
\arguments{
\item{pred}{predicted probabilities}

\item{y}{response}

\item{col}{color}

\item{plot.values}{wether the value should be plotted}

\item{...}{passed to abline}
}
\description{
Adds Prevalence on RBP curve
}
\author{
Giuseppe Casalicchio
}

