% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeAxis}
\alias{computeAxis}
\title{Function computes x and y Axis for RBP plot}
\usage{
computeAxis(pred, y)
}
\arguments{
\item{pred}{predicted probabilities}

\item{y}{response}
}
\description{
Function computes x and y Axis for RBP plot
}
\author{
Giuseppe Casalicchio
}

