% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotRBPcurve}
\alias{plotRBPcurve}
\title{Plot residual-based predictiveness curve.}
\usage{
plotRBPcurve(obj, main = "RBP Curve", xlab = "Cumulative Percentage",
  ylab = "Estimated Residuals", type = "l", ylim = c(-1, 1),
  conditionalAxis = FALSE, title.line = ifelse(conditionalAxis, 3, 2),
  add = FALSE, ...)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{main}{plot title}

\item{xlab}{}

\item{ylab}{label for y-axis}

\item{type}{plot type}

\item{ylim}{limits for y-axis}

\item{title.line}{where to plot the title}

\item{add}{[\code{logical(1)}]\cr
Should RBP plot be added to current plot?
Default is \code{FALSE}.}

\item{...}{passed to plot or lines (depending on add)}

\item{xlab}{[\code{character(1)}]\cr
Label for X-axis.
Default is \dQuote{Cumulative Percentage}.}

\item{ylab}{[\code{character(1)}]\cr
Label for Y-axis.
Default is \dQuote{Estimated Residuals}.}

\item{type}{[\code{logical(1)}]\cr
Should an additional axis be plotted reflecting residuals conditional on y?
Default is \code{FALSE}.}

\item{cond.axis}{[\code{logical(1)}]\cr
Should an additional axis be plotted reflecting residuals conditional on y?
Default is \code{FALSE}.}
}
\description{
Plot residual-based predictiveness curve.
}

