% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RBPcurve}
\alias{RBPcurve}
\title{RBP plot}
\usage{
RBPcurve(pred, y, main="RBP Curve",
                xlab="Cumulative Percentage", ylab="Estimated Residuals",
                type="l", ylim=c(-1,1), conditionalAxis=FALSE,
                title.line=ifelse(conditionalAxis, 3, 2), add=FALSE, ...)
}
\arguments{
\item{pred}{predicted probabilities}

\item{y}{response}

\item{main}{plot title}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{type}{plot type}

\item{ylim}{limits for y-axis}

\item{conditionalAxis}{wether an additional axis should be plotted reflecting residuals conditional on y}

\item{title.line}{where to plot the title}

\item{add}{wether RBP plot should be added to current plot}

\item{...}{passed to plot or lines (depending on add)}
}
\description{
Plot for Residual-Based Predictiveness Curve
}
\author{
Giuseppe Casalicchio
}

