% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addGoodCalib}
\alias{addGoodCalib}
\title{Displays integrals that refer to good calibration.}
\usage{
addGoodCalib(obj, plot.values = TRUE)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{plot.values}{[\code{logical(1)}]\cr
Should values be added to plot?
Default is \code{FALSE}.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
Displays integrals that refer to good calibration.
}

