% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addPEV}
\alias{addPEV}
\title{PEV}
\usage{
addPEV(pred, y, plot.values=TRUE, ...)
}
\arguments{
\item{pred}{predicted probabilities}

\item{y}{response}

\item{plot.values}{wether the value should be plotted}

\item{...}{passed to abline}
}
\description{
Highlights Area that is used to compute the PEV
}
\author{
Giuseppe Casalicchio
}

