% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addPrevalence}
\alias{addPrevalence}
\title{Add prevalence to RBP curve.}
\usage{
addPrevalence(obj, col = "gray", plot.values = TRUE, ...)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{col}{color}

\item{plot.values}{[\code{logical(1)}]\cr
Should values be added to plot?
Default is \code{FALSE}.}

\item{...}{passed to abline}
}
\value{
[\code{invisible(NULL)}].
}
\description{
Add prevalence to RBP curve.
}

