% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addPrevalence.R
\name{addPrevalence}
\alias{addPrevalence}
\title{Visualizes the prevalence on the RBP curve.}
\usage{
addPrevalence(obj, plot.values = TRUE, digits = 3L, col = "grey")
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{plot.values}{[\code{logical(1)}]\cr
Whether the values of the corresponding measure should be added to the plot?
Default is \code{FALSE}.}

\item{digits}{[\code{numeric(1)}]\cr
Indicates the number of decimal places for the values that are plotted when
\code{plot.values = TRUE}.
Default is \code{3L}.}

\item{col}{[\code{character(1)} | \code{numeric(1)}]\cr
A specification for the plotting color.}
}
\value{
[\code{invisible(NULL)}].
}
\description{
The prevalence is the proportion of a population having a specific condition.
In binary classification, the condition refers to whether the target variable has the value
\code{1}, that is, whether the target variable corresponds to the positive class.
}

