% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addRates}
\alias{addRates}
\title{Visualizes the TPR and FPR on the RBP curve.}
\usage{
addRates(obj, plot.values = TRUE, t = obj$prevalence, col = "grey", ...)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{plot.values}{[\code{logical(1)}]\cr
Whether the values of the corresponding measure should be added to the plot?
Default is \code{FALSE}.}

\item{t}{Threshold that is used to compute the true positve and false positive rate.}

\item{col}{[\code{character}(1) | \code{numeric}(1) ]\cr
A specification for the plotting color.}

\item{...}{currently not used}
}
\description{
For a given threshold \code{t}, the true positive rate (TPR) and the false positive
rate (FPR) can be visually assessed by the RBP curve by the intersection of the RBP curve with
the horizontal lines at \code{-t} and \code{1-t}, respectively.
}

