% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addRates}
\alias{addRates}
\title{Add tpr and fpr on RBP curve.}
\usage{
addRates(obj, plot.values = TRUE, ...)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{plot.values}{[\code{logical(1)}]\cr
Should values be added to plot?
Default is \code{FALSE}.}

\item{...}{currently not used}
}
\description{
Add tpr and fpr on RBP curve.
}

