% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addRates}
\alias{addRates}
\title{TPR and FPR}
\usage{
addRates(pred, y, plot.values=TRUE, ...)
}
\arguments{
\item{pred}{predicted probabilities}

\item{y}{response}

\item{plot.values}{wether the value should be plotted}

\item{...}{currently not used}
}
\description{
Adds TPR and FPR on RBP curve
}
\author{
Giuseppe Casalicchio
}

