% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addWellCalib}
\alias{addWellCalib}
\title{Add TPR and FPR on RBP curve.}
\usage{
addWellCalib(obj, plot.values = TRUE)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{plot.values}{[\code{logical(1)}]\cr
Should values be added to plot?
Default is \code{FALSE}.}
}
\description{
Add TPR and FPR on RBP curve.
}

