% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotRBPcurve}
\alias{plotRBPcurve}
\title{Plot residual-based predictiveness (RBP) curve.}
\usage{
plotRBPcurve(obj, main = "RBP Curve", xlab = "Cumulative Percentage",
  ylab = "Estimated Residuals", type = "l", ylim = c(-1, 1.1),
  cond.axis = FALSE, title.line = ifelse(cond.axis, 3, 2), add = FALSE,
  ...)
}
\arguments{
\item{obj}{[\code{\link{RBPObj}}]\cr
Data container for RBP curve.}

\item{main}{An overall title for the plot.}

\item{xlab}{[\code{character(1)}]\cr
Label for X-axis.
Default is \dQuote{Cumulative Percentage}.}

\item{ylab}{[\code{character(1)}]\cr
Label for Y-axis.
Default is \dQuote{Estimated Residuals}.}

\item{type}{[\code{character(1)}]\cr
The plot type that should be drawn, see \code{\link{plot}} for all possible types.
Default is \code{type = "l"} for \bold{l}ines.}

\item{ylim}{limits for Y-axis. Default is \code{ylim = c(-1,1.1)}.}

\item{cond.axis}{[\code{logical(1)}]\cr
Should an additional axis be plotted reflecting residuals conditional on y?
Default is \code{FALSE}.}

\item{title.line}{where to plot the title, see \code{\link{title}}.}

\item{add}{[\code{logical(1)}]\cr
Should RBP plot be added to current plot?
Default is \code{FALSE}.}

\item{...}{passed to plot or lines (depending on add)}
}
\description{
Plot residual-based predictiveness (RBP) curve.
}

